pyproject.toml
[project]
name = "servicenow-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.6",
    "langgraph-cli[inmem]>=0.4.19",
    "python-dotenv>=1.2.2",
]
